[Add] 필터로 병원검색 구현 거리 5km, 진료과목까지 필터 완료 크롤링 후 운영시간 필터 추가 필요

+++ back/src/main/java/com/roller/doc/api/response/hospital/HospitalDescRes.java
@@ -7,13 +7,13 @@ import java.util.List;
 
 @Data
 public class HospitalDescRes {
-    private int hospitalId;
+    private long hospitalId;
     private String hospitalAdd; //주소
     private int hospitalParking; //주차 대수
     private String hospitalDevice; //의료장비
     private String hospitalSpecial; //특수진료
     @Builder
-    public HospitalDescRes(int hospitalId, String hospitalAdd, int hospitalParking, String hospitalDevice,
+    public HospitalDescRes(long hospitalId, String hospitalAdd, int hospitalParking, String hospitalDevice,
                            String hospitalSpecial) {
         this.hospitalId = hospitalId;
         this.hospitalAdd = hospitalAdd;
